// src/interfaces/position.interface.ts
import { Student } from './student.interface';
export interface Position {
  position_id?: number;
  chiefJustice: string;
  associateJustice: string;
  speakerHouse: string;
  internalDeputy: string;
  externalDeputy: string;
  ministerHealth: string;
  ministerInfo: string;
  ministerPlanning: string;
  ministerAcadAffairs: string;
  ministerFinance: string;
  commissionAudit: string;
  commissionElection: string;
  commissionWelfare: string;
}